feat: moved vestings card into one component
