+++ features/claim/vestingCarousel.tsx
@@ -1,5 +1,4 @@
-import { useVestingsContext } from 'features/vesting';
-import { VestingCard } from './vestingCard';
+import { VestingCardDetailed, useVestingsContext } from 'features/vesting';
 import { useCallback, useMemo } from 'react';
 import Swiper from 'swiper';
 import { Carousel } from 'shared/ui';
@@ -23,7 +22,11 @@ export const VestingCarousel = () => {
   return (
     <Carousel onSlideChange={handleSlideChange}>
       {vestingsView?.map((vesting, index) => (
-        <VestingCard key={vesting.escrow} vesting={vesting} index={index} />
+        <VestingCardDetailed
+          key={vesting.escrow}
+          vesting={vesting}
+          index={index}
+        />
       ))}
     </Carousel>
   );
