+++ features/vesting/index.ts
@@ -1,3 +1,5 @@
 export * from './hooks';
 export * from './vestingsContext';
 export * from './types';
+export * from './vestingCardDetailed';
+export * from './vestingCardSummary';
